backend-data-process/requirements.txt
apache-airflow==2.9.3
evaluate==0.4.1
groq==0.5.0
langchain==0.1.16
langchain_groq==0.1.3
langchain-openai==0.1.4
langid==1.1.6
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.23.6
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
scikit-learn==1.4.2
selenium==4.20.0
chromedriver-py==126.0.6478.182
seaborn==0.13.2
stem==1.8.2
SQLAlchemy==1.4.*
sqlparse==0.5.0
swifter==1.4.0
tqdm==4.66.2
unidecode==1.3.8
yfinance==0.2.38
